Fix listing related eslint PropTypes violations
